python-code-review/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
boto3==1.12.28            # via -r requirements.in
botocore==1.15.28         # via boto3, s3transfer
docutils==0.15.2          # via botocore
jmespath==0.9.5           # via boto3, botocore
python-dateutil==2.8.1    # via botocore
pyyaml==5.3.1             # via -r requirements.in
s3transfer==0.3.3         # via boto3
six==1.14.0               # via python-dateutil
urllib3==1.25.8           # via botocore
